IoT/requirements.txt
opencv-python==4.11.0.86
numpy==2.2.0
pillow==11.1.0
fastapi==0.115.4
uvicorn==0.30.1
python-multipart==0.0.9
deepface==0.0.93